Build just once weekly
Reduce impact on actions billing by reducing full-scale infra build to once a week. This is enough to maintain security patching for low, moderate and high priority vulnerabilities in a reasonable timescale. Critical patching will be done manually based on CVE alerting.

Also weekly is enough to maintain high resolution of the decay of software dependencies out in the wild, especially in python.